requirements-runtime.txt
fastapi>=0.135
uvicorn>=0.44
pydantic>=2.12
reportlab>=4.4
jinja2>=3.1
pyyaml>=6.0
python-multipart>=0.0.9
python-docx>=1.1
google-genai>=1.0
